+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Improve logging in `ReplicantSessionManagerImpl` so that long expand duration log messages are auto-promoted to `SEVERE` log level.
+* Fix incoming counts in log levels.
+
 ### [v6.171](https://github.com/replicant4j/replicant/tree/v6.171) (2026-02-03) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.170...v6.171)
 
 Changes in this release: